# backend-java/agent-api/src/main/resources/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Agent Ops API",
    "description": "Agent Ops 后端接口文档，可直接导入 Apifox 管理接口契约。",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地 Java 后端"
    }
  ],
...
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["ExternalAuth"],
        "summary": "用户注册",
        "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } } },
        "responses": {
          "201": { "description": "注册成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["ExternalAuth"],
        "summary": "用户登录",
        "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } } },
        "responses": {
          "200": { "description": "登录成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/chat/messages": {
      "post": {
        "tags": ["ExternalChat"],
        "summary": "提交问题并获取 RAG 回答",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatRequest" } } } },
        "responses": {
          "200": { "description": "回答结果", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/chat/sessions": {
      "get": {
        "tags": ["ExternalChat"],
        "summary": "查询会话列表",
        "parameters": [
          { "name": "kbId", "in": "query", "schema": { "type": "integer", "format": "int64" } },
          { "name": "keyword", "in": "query", "schema": { "type": "string" } },
          { "name": "startDate", "in": "query", "schema": { "type": "string" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string" } },
          { "name": "filterCustomerId", "in": "query", "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "会话列表", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ChatSessionResponse" } } } } }
        }
      }
    },
    "/api/chat/sessions/{sessionId}/messages": {
      "get": {
        "tags": ["ExternalChat"],
        "summary": "查询会话消息明细",
        "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
        "responses": {
          "200": { "description": "消息列表", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ChatMessageResponse" } } } } }
        }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "tags": ["ExternalAdminUser"],
        "summary": "查询用户列表",
        "parameters": [{ "name": "customerId", "in": "query", "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "用户列表", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/UserManageResponse" } } } } }
        }
      },
      "post": {
        "tags": ["ExternalAdminUser"],
        "summary": "创建用户",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserManageRequest" } } } },
        "responses": {
          "201": { "description": "创建成功", "content": { "application/json": { "schema": { "type": "integer", "format": "int64" } } } }
        }
      }
    },
    "/api/admin/users/{id}": {
      "put": {
        "tags": ["ExternalAdminUser"],
        "summary": "更新用户",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserManageRequest" } } } },
        "responses": { "200": { "description": "更新成功" } }
      }
    },
    "/api/admin/users/{id}/status": {
      "patch": {
        "tags": ["ExternalAdminUser"],
        "summary": "启用或禁用用户",
        "parameters": [
          { "$ref": "#/components/parameters/Id" },
          { "name": "enabled", "in": "query", "required": true, "schema": { "type": "boolean" } }
        ],
        "responses": { "200": { "description": "状态已更新" } }
      }
    },
    "/api/admin/customers": {
      "get": {
        "tags": ["ExternalCustomer"],
        "summary": "查询客户空间列表",
        "responses": {
          "200": { "description": "客户列表", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/CustomerResponse" } } } } }
        }
      },
      "post": {
        "tags": ["ExternalCustomer"],
        "summary": "创建客户空间",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CustomerRequest" } } } },
        "responses": {
          "201": { "description": "创建成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CustomerResponse" } } } }
        }
      }
    },
    "/api/admin/customers/{id}": {
      "get": {
        "tags": ["ExternalCustomer"],
        "summary": "查询客户详情",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "客户详情", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CustomerResponse" } } } }
        }
      },
      "put": {
        "tags": ["ExternalCustomer"],
        "summary": "更新客户空间",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CustomerRequest" } } } },
        "responses": { "200": { "description": "更新成功" } }
      },
      "delete": {
        "tags": ["ExternalCustomer"],
        "summary": "删除客户空间",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": { "204": { "description": "删除成功" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
...
      "SessionId": { "name": "sessionId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
...
    },
    "responses": {
      "BadRequest": { "description": "请求参数错误", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
      "Unauthorized": { "description": "未认证或 Token 失效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
...
    },
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "path": { "type": "string" },
          "traceId": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password", "displayName", "inviteCode"],
        "properties": {
          "username": { "type": "string", "minLength": 2, "maxLength": 128 },
          "password": { "type": "string", "minLength": 6, "maxLength": 128, "format": "password" },
          "displayName": { "type": "string", "maxLength": 128 },
          "email": { "type": "string", "format": "email" },
          "inviteCode": { "type": "string", "maxLength": 64 }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "minLength": 2, "maxLength": 128 },
          "password": { "type": "string", "minLength": 6, "maxLength": 128, "format": "password" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "tokenType": { "type": "string", "example": "Bearer" },
          "user": { "$ref": "#/components/schemas/UserInfo" }
        }
      },
      "UserInfo": {
        "type": "object",
        "properties": {
          "userId": { "type": "integer", "format": "int64" },
          "customerId": { "type": "integer", "format": "int64" },
          "username": { "type": "string" },
          "displayName": { "type": "string" },
          "role": { "type": "string" }
        }
      },
      "ChatRequest": {
        "type": "object",
        "required": ["question"],
        "properties": {
          "sessionId": { "type": "integer", "format": "int64", "nullable": true },
          "kbId": { "type": "integer", "format": "int64", "nullable": true },
          "question": { "type": "string" }
        }
      },
      "Citation": {
        "type": "object",
        "properties": {
          "chunk_id": { "type": "string" },
          "doc_id": { "type": "string" },
          "title": { "type": "string" },
          "score": { "type": "number", "format": "double", "nullable": true },
          "text": { "type": "string" }
        }
      },
      "ChatResponse": {
        "type": "object",
        "properties": {
          "sessionId": { "type": "integer", "format": "int64" },
          "traceId": { "type": "string" },
          "answer": { "type": "string" },
          "citations": { "type": "array", "items": { "$ref": "#/components/schemas/Citation" } }
        }
      },
      "ChatSessionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "customerId": { "type": "integer", "format": "int64" },
          "userId": { "type": "integer", "format": "int64" },
          "kbId": { "type": "integer", "format": "int64", "nullable": true },
          "title": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ChatMessageResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "customerId": { "type": "integer", "format": "int64" },
          "sessionId": { "type": "integer", "format": "int64" },
          "role": { "type": "string", "enum": ["user", "assistant", "system"] },
          "content": { "type": "string" },
          "traceId": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "UserManageRequest": {
        "type": "object",
        "required": ["username", "displayName", "roleCode"],
        "properties": {
          "username": { "type": "string", "minLength": 2, "maxLength": 128 },
          "password": { "type": "string", "minLength": 6, "maxLength": 128, "format": "password" },
          "displayName": { "type": "string", "maxLength": 128 },
          "email": { "type": "string" },
          "roleCode": { "type": "string" },
          "customerId": { "type": "integer", "format": "int64" },
          "status": { "type": "string" }
        }
      },
      "UserManageResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "customerId": { "type": "integer", "format": "int64" },
          "username": { "type": "string" },
          "displayName": { "type": "string" },
          "email": { "type": "string" },
          "roleCode": { "type": "string" },
          "status": { "type": "string" },
          "createdAt": { "type": "string" }
        }
      },
      "CustomerRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "maxLength": 128 },
          "contactName": { "type": "string", "maxLength": 128 },
          "contactEmail": { "type": "string", "maxLength": 256 },
          "status": { "type": "string" }
        }
      },
      "CustomerResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "contactName": { "type": "string" },
          "contactEmail": { "type": "string" },
          "inviteCode": { "type": "string" },
          "status": { "type": "string" },
          "createdAt": { "type": "string" }
        }
      },
...
    }
  }
}